backend/requirements.txt
asgiref==3.8.1
Django==5.1.8
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.1.0
sqlparse==0.5.3
cryptography>=43.0.0
